+++ src/sinapse/component.clj
@@ -1,11 +1,18 @@
 (ns sinapse.component
-  (:require [integrant.core :as ig]
+  (:require [iapetos.core :as prometheus]
+            [integrant.core :as ig]
             [clojure.tools.logging :as log]
             [io.pedestal.interceptor :as interceptor]
             [io.pedestal.interceptor.chain :as chain]
             [common-clj.traceability.core :as traceability]
+            [medley.core :as medley]
             [schema.core :as s])
   (:import (clojure.lang IFn PersistentQueue)))
+
+(def metrics
+  [(prometheus/counter :produced-messages {:labels [:service :topic :host]})
+   (prometheus/counter :consumed-messages {:labels [:service :topic :host]})
+   (prometheus/counter :replayed-messages {:labels [:service :topic :host]})])
 
 (s/defschema Consumers
   {s/Str {:schema       s/Any
@@ -22,9 +29,12 @@
 
 (defn enqueue!
   [{:keys [topic] :as message}
-   sinapse]
+   {:keys [registry service-name] :as sinapse}]
   (let [message' (assoc message :meta {:correlation-id (traceability/current-correlation-id!)})]
-    (swap! (get-in sinapse [:messages topic]) conj message')))
+    (swap! (get-in sinapse [:messages topic]) conj message')
+    (when registry
+      (prometheus/inc registry :produced-messages {:service service-name
+                                                   :topic   topic}))))
 
 (defmethod ig/init-key ::sinapse
   [_ {:keys [components consumers]}]
@@ -32,10 +42,13 @@
 
   (s/validate Consumers consumers)
 
-  (let [bag-of-messages (reduce (fn [acc current]
+  (let [prometheus (:prometheus components)
+        bag-of-messages (reduce (fn [acc current]
                                   (assoc acc (:title current) (atom PersistentQueue/EMPTY)))
                                 {} (-> components :config :topics))
-        sinapse {:messages bag-of-messages}]
+        sinapse (medley/assoc-some {:messages     bag-of-messages
+                                    :service-name (-> components :config :service-name)}
+                                   :registry (:registry prometheus))]
 
     (doseq [{:keys [title parallel-consumers]} (-> components :config :topics)
             :let [schema (get-in consumers [title :schema])
@@ -47,14 +60,21 @@
                   (let [[old _new] (-> (get bag-of-messages title)
                                        (swap-vals! pop))
                         {:keys [payload] :as message} (peek old)]
-                    (try
-                      (s/validate schema payload)
-                      (chain/execute {:payload    payload
-                                      :components (assoc components :sinapse sinapse)}
-                                     (conj (or interceptors []) (handler-fn->interceptor handler-fn)))
-                      (catch Exception ex
-                        (log/error ::error-while-consuming-message :exception ex)
-                        (enqueue! message sinapse))))))))
+                    (when message
+                      (try
+                        (s/validate schema payload)
+                        (chain/execute {:payload    payload
+                                        :components (assoc components :sinapse sinapse)}
+                                       (conj (or interceptors []) (handler-fn->interceptor handler-fn)))
+                        (when prometheus
+                          (prometheus/inc (:registry prometheus) :consumed-messages {:service (-> components :config :service-name)
+                                                                                     :topic   title}))
+                        (catch Exception ex
+                          (log/error ::error-while-consuming-message :exception ex)
+                          (enqueue! message sinapse)
+                          (when prometheus
+                            (prometheus/inc (:registry prometheus) :replayed-messages {:service (-> components :config :service-name)
+                                                                                       :topic   title}))))))))))
     sinapse))
 
 (defmethod ig/halt-key! ::sinapse