+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions
 of [keepachangelog.com](http://keepachangelog.com/).
 
+## 0.1.0-4 - 2025-03-03 (beta)
+
+### Added
+
+- Adding prometheus metrics for produced, consumed and failed (replayed) messages.
+
+## 0.1.0-3 - 2025-03-03 (beta)
+
+### Fixed
+
+- Fixed component initialization issue, where there is no message when the component is initialized.
+
 ## 0.1.0-2 - 2025-03-03 (beta)
 
 ### Fixed